+++ tests/unit/processor/pseudonymizer/test_pseudonymizer.py
@@ -24,7 +24,7 @@ TLD_LIST = f"file://{Path().absolute().joinpath(REL_TLD_LIST_PATH).as_posix()}"
 class TestPseudonymizer(BaseProcessorTestCase):
     CONFIG = {
         "type": "pseudonymizer",
-        "pseudonyms_topic": "pseudonyms",
+        "outputs": [{"kafka": "topic"}],
         "pubkey_analyst": "tests/testdata/unit/pseudonymizer/example_analyst_pub.pem",
         "pubkey_depseudo": "tests/testdata/unit/pseudonymizer/example_depseudo_pub.pem",
         "hash_salt": "a_secret_tasty_ingredient",
@@ -38,6 +38,37 @@ class TestPseudonymizer(BaseProcessorTestCase):
     def setup_method(self) -> None:
         super().setup_method()
         self.regex_mapping = self.CONFIG.get("regex_mapping")
+
+    @pytest.mark.parametrize(
+        "config_change, error, msg",
+        [
+            ({"outputs": [{"kafka": "topic"}]}, None, None),
+            ({"outputs": []}, ValueError, "Length of 'outputs' must be => 1"),
+            (
+                {"outputs": [{"kafka": 1}]},
+                TypeError,
+                "must be <class 'str'>",
+            ),
+            (
+                {"outputs": [{1: "topic"}]},
+                TypeError,
+                "must be <class 'str'>",
+            ),
+            (
+                {"outputs": [{"kafka": "topic", "opensearch": "index_1"}]},
+                ValueError,
+                "Length of 'outputs' must be <= 1",
+            ),
+        ],
+    )
+    def test_config_validation(self, config_change, error, msg):
+        config = deepcopy(self.CONFIG)
+        config |= config_change
+        if error:
+            with pytest.raises(error, match=msg):
+                Factory.create({"name": config}, self.logger)
+        else:
+            Factory.create({"name": config}, self.logger)
 
     def test_pseudonymize_event(self):
         event_raw = {"foo": "bar"}