+++ tests/unit/processor/selective_extractor/test_selective_extractor.py
@@ -36,10 +36,8 @@ class TestSelectiveExtractor(BaseProcessorTestCase):
             {
                 "filter": field_name,
                 "selective_extractor": {
-                    "extract": {
-                        "extracted_field_list": [field_name],
-                        "target_topic": "my topic",
-                    },
+                    "source_fields": [field_name],
+                    "outputs": [{"kafka": "topic"}],
                 },
             }
         )
@@ -52,32 +50,48 @@ class TestSelectiveExtractor(BaseProcessorTestCase):
         else:
             assert False
 
-    def test_process_returns_selective_extractor_topic(self):
+    def test_process_returns_selective_extractor_target_topic(self):
         field_name = f"{uuid.uuid4()}"
-        rule = SelectiveExtractorRule._create_from_dict(
-            {
-                "filter": field_name,
-                "selective_extractor": {
-                    "extract": {
-                        "extracted_field_list": [field_name],
-                        "target_topic": "my topic",
-                    },
-                },
-            }
-        )
-        self.object._specific_tree.add_rule(rule)
+        rule = {
+            "filter": field_name,
+            "selective_extractor": {
+                "source_fields": [field_name],
+                "outputs": [{"opensearch": "my topic"}],
+            },
+        }
+        self._load_specific_rule(rule)
         document = {field_name: "test_message", "other": "field"}
         result = self.object.process(document)
-        for _, topic in result:
-            if topic == "my topic":
-                break
-        else:
-            assert False
+        output = result[0][1][0]
+        assert "my topic" in output.values()
+
+    def test_process_returns_selective_extractor_target_output(self):
+        field_name = f"{uuid.uuid4()}"
+        rule = {
+            "filter": field_name,
+            "selective_extractor": {
+                "source_fields": [field_name],
+                "outputs": [{"opensearch": "index"}],
+            },
+        }
+        self._load_specific_rule(rule)
+        document = {field_name: "test_message", "other": "field"}
+        result = self.object.process(document)
+        output = result[0][1][0]
+        assert "opensearch" in output.keys()
 
     def test_process_returns_extracted_fields(self):
         document = {"message": "test_message", "other": "field"}
+        rule = {
+            "filter": "message",
+            "selective_extractor": {
+                "source_fields": ["message"],
+                "outputs": [{"opensearch": "index"}],
+            },
+        }
+        self._load_specific_rule(rule)
         result = self.object.process(document)
-        for filtered_event, _ in result:
+        for filtered_event, *_ in result:
             if filtered_event[0] == {"message": "test_message"}:
                 break
         else:
@@ -104,31 +118,27 @@ class TestSelectiveExtractor(BaseProcessorTestCase):
             mock_apply_rules.assert_called()
 
     def test_process_extracts_dotted_fields(self):
-        rule = SelectiveExtractorRule._create_from_dict(
-            {
-                "filter": "message",
-                "selective_extractor": {
-                    "extract": {
-                        "extracted_field_list": ["other.message", "message"],
-                        "target_topic": "my topic",
-                    },
-                },
-            }
-        )
-        self.object._specific_tree.add_rule(rule)
+        rule = {
+            "filter": "message",
+            "selective_extractor": {
+                "source_fields": ["other.message", "message"],
+                "outputs": [{"opensearch": "index"}],
+            },
+        }
+        self._load_specific_rule(rule)
         document = {"message": "test_message", "other": {"message": "my message value"}}
         result = self.object.process(document)
 
-        for extracted_event, _ in result:
+        for extracted_event, *_ in result:
             if extracted_event[0].get("other", {}).get("message") is not None:
                 break
         else:
             assert False, f"other.message not in {result}"
 
     def test_process_clears_internal_filtered_events_list_before_every_event(self):
-        assert len(self.object._filtered_events) == 0
+        assert len(self.object._extra_data) == 0
         document = {"message": "test_message", "other": {"message": "my message value"}}
         _ = self.object.process(document)
-        assert len(self.object._filtered_events) == 1
+        assert len(self.object._extra_data) == 1
         _ = self.object.process(document)
-        assert len(self.object._filtered_events) == 1
+        assert len(self.object._extra_data) == 1